Add post processing steps for yolov5,v8,v9,v10 and fix minor bugs (#99)
### Ticket

- https://github.com/tenstorrent/tt-forge/issues/374
- Fixes https://github.com/tenstorrent/tt-forge-models/issues/94

### Problem description

- YOLOv5, YOLOv8, YOLOv9, and YOLOv10 are missing post-processing steps.
- YOLOv5 also has incorrect preprocessing steps.
- ResNet post-processing includes an unnecessary BFP16 conversion of
inputs.

### What's changed

- Added the required post-processing steps for YOLOv5, YOLOv8, YOLOv9,
and YOLOv10.
- Fixed YOLOv5 preprocessing.
- Removed the unnecessary BFP16 input conversion in ResNet
post-processing.

### Checklist
- [x] verified the changes with local testing

### Logs

-
[set2_check.zip](https://github.com/user-attachments/files/21896993/set2_check.zip)

+++ resnet/pytorch/src/utils.py
@@ -21,10 +21,7 @@ def run_and_print_results(framework_model, compiled_model, inputs):
 
     results = []
     for i, image in enumerate(inputs):
-        processed_inputs = processor(image, return_tensors="pt")["pixel_values"].to(
-            torch.bfloat16
-        )
-
+        processed_inputs = processor(image, return_tensors="pt")["pixel_values"]
         cpu_logits = framework_model(processed_inputs)[0]
         cpu_conf, cpu_idx = cpu_logits.softmax(-1).max(-1)
         cpu_pred = label_dict.get(cpu_idx.item(), "Unknown")
